requirements.txt
python-actionlib
python-rospy
yolov4
cvbridge3
opencv-python
python-smach
smach
roslibpy
smach_ros